+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2025-07-30
+
+### Added
+
+- Added a minimum size for the application
+- Added a message to the user when the terminal is too small
+
 ## [0.3.1] - 2025-07-30
 
 ### Changed